# backend/openapi.json
{
  "components": {
    "schemas": {
      "ActivityFrequencyCode": {
        "enum": [
          "less_monthly",
          "monthly",
          "two_three_monthly",
          "weekly",
          "two_three_weekly",
          "four_plus_weekly",
          "irregular"
        ],
        "title": "ActivityFrequencyCode",
        "type": "string"
      },
      "ActivityLocationRead": {
        "additionalProperties": false,
        "properties": {
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "facilityName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Facilityname"
          },
          "isOnline": {
            "title": "Isonline",
            "type": "boolean"
          },
          "nearestStation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Neareststation"
          },
          "prefecture": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prefecture"
          },
          "publicNote": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publicnote"
          }
        },
        "required": [
          "prefecture",
          "city",
          "facilityName",
          "nearestStation",
          "isOnline",
          "publicNote"
        ],
        "title": "ActivityLocationRead",
        "type": "object"
      },
      "ActivityScheduleRead": {
        "additionalProperties": false,
        "properties": {
          "endsAt": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Endsat"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "startsAt": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startsat"
          },
          "timeBand": {
            "$ref": "#/components/schemas/TimeBand"
          },
          "weekday": {
            "$ref": "#/components/schemas/WeekdayValue"
          }
        },
        "required": [
          "weekday",
          "timeBand",
          "startsAt",
          "endsAt",
          "note"
        ],
        "title": "ActivityScheduleRead",
        "type": "object"
      },
...
      "CampFrequencyCode": {
        "enum": [
          "none",
          "once_year",
          "twice_year",
          "three_plus_year",
          "unknown"
        ],
        "title": "CampFrequencyCode",
        "type": "string"
      },
      "CategoryRead": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug"
        ],
        "title": "CategoryRead",
        "type": "object"
      },
      "CircleCollectionResponse": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/CircleListItem"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/RequestMetadata"
          },
          "page": {
            "$ref": "#/components/schemas/CirclePageMetadata"
          }
        },
        "required": [
          "data",
          "page",
          "meta"
        ],
        "title": "CircleCollectionResponse",
        "type": "object"
      },
...
      "CircleListItem": {
        "additionalProperties": false,
        "properties": {
          "activityLocations": {
            "items": {
              "$ref": "#/components/schemas/ActivityLocationRead"
            },
            "title": "Activitylocations",
            "type": "array"
          },
          "activitySchedules": {
            "items": {
              "$ref": "#/components/schemas/ActivityScheduleRead"
            },
            "title": "Activityschedules",
            "type": "array"
          },
          "category": {
            "$ref": "#/components/schemas/CategoryRead"
          },
          "circleType": {
            "$ref": "#/components/schemas/CircleType"
          },
          "displayName": {
            "title": "Displayname",
            "type": "string"
          },
          "featuredTags": {
            "items": {
              "$ref": "#/components/schemas/FeaturedTagRead"
            },
            "maxItems": 5,
            "title": "Featuredtags",
            "type": "array"
          },
          "headline": {
            "title": "Headline",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "officialStatus": {
            "$ref": "#/components/schemas/OfficialStatus"
          },
          "publishedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publishedat"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "universities": {
            "items": {
              "$ref": "#/components/schemas/UniversityRead"
            },
            "title": "Universities",
            "type": "array"
          }
        },
        "required": [
          "id",
          "displayName",
          "headline",
          "summary",
          "officialStatus",
          "circleType",
          "publishedAt",
          "category",
          "universities",
          "featuredTags",
          "activitySchedules",
          "activityLocations"
        ],
        "title": "CircleListItem",
        "type": "object"
      },
      "CirclePageMetadata": {
        "additionalProperties": false,
        "properties": {
          "hasMore": {
            "title": "Hasmore",
            "type": "boolean"
          },
          "limit": {
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "nextCursor": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          },
          "totalCount": {
            "minimum": 0.0,
            "title": "Totalcount",
            "type": "integer"
          }
        },
        "required": [
          "hasMore",
          "limit",
          "totalCount"
        ],
        "title": "CirclePageMetadata",
        "type": "object"
      },
      "CircleSort": {
        "enum": [
          "newest"
        ],
        "title": "CircleSort",
        "type": "string"
      },
      "CircleType": {
        "enum": [
          "circle",
          "club",
          "intercollegiate",
          "student_organization"
        ],
        "title": "CircleType",
        "type": "string"
      },
...
      "FeaturedTagRead": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug"
        ],
        "title": "FeaturedTagRead",
        "type": "object"
      },
      "GenderBalanceCode": {
        "enum": [
          "women_majority",
          "balanced",
          "men_majority",
          "mixed_or_other",
          "not_disclosed"
        ],
        "title": "GenderBalanceCode",
        "type": "string"
      },
      "MemberCountBand": {
        "enum": [
          "1_10",
          "11_30",
          "31_80",
          "81_150",
          "151_plus",
          "not_disclosed"
        ],
        "title": "MemberCountBand",
        "type": "string"
      },
      "OfficialStatus": {
        "enum": [
          "official",
          "unofficial",
          "unknown"
        ],
        "title": "OfficialStatus",
        "type": "string"
      },
...
      "RelationshipType": {
        "enum": [
          "primary",
          "participating",
          "activity_base"
        ],
        "title": "RelationshipType",
        "type": "string"
      },
      "RequestMetadata": {
        "additionalProperties": false,
        "properties": {
          "requestId": {
            "format": "uuid",
            "title": "Requestid",
            "type": "string"
          }
        },
        "required": [
          "requestId"
        ],
        "title": "RequestMetadata",
        "type": "object"
      },
...
      "TimeBand": {
        "enum": [
          "morning",
          "daytime",
          "evening",
          "night",
          "all_day",
          "irregular"
        ],
        "title": "TimeBand",
        "type": "string"
      },
      "UniversityRead": {
        "additionalProperties": false,
        "properties": {
          "campus": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Campus"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "relationshipType": {
            "$ref": "#/components/schemas/RelationshipType"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "relationshipType",
          "campus"
        ],
        "title": "UniversityRead",
        "type": "object"
      },
      "WeekdayValue": {
        "enum": [
          "1",
          "2",
          "3",
          "4",
          "5",
          "6",
          "7",
          "irregular"
        ],
        "title": "WeekdayValue",
        "type": "string"
      }
    }
  },
  "info": {
    "title": "Campus Circle API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/circles": {
      "get": {
        "operationId": "listPublicCircles",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "officialStatus",
            "required": false,
            "schema": {
              "items": {
                "$ref": "#/components/schemas/OfficialStatus"
              },
              "title": "Officialstatus",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "circleType",
            "required": false,
            "schema": {
              "items": {
                "$ref": "#/components/schemas/CircleType"
              },
              "title": "Circletype",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "campFrequencyCode",
            "required": false,
            "schema": {
              "items": {
                "$ref": "#/components/schemas/CampFrequencyCode"
              },
              "title": "Campfrequencycode",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "memberCountBand",
            "required": false,
            "schema": {
              "items": {
                "$ref": "#/components/schemas/MemberCountBand"
              },
              "title": "Membercountband",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "activityFrequencyCode",
            "required": false,
            "schema": {
              "items": {
                "$ref": "#/components/schemas/ActivityFrequencyCode"
              },
              "title": "Activityfrequencycode",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "weekday",
            "required": false,
            "schema": {
              "items": {
                "$ref": "#/components/schemas/WeekdayValue"
              },
              "title": "Weekday",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "timeBand",
            "required": false,
            "schema": {
              "items": {
                "$ref": "#/components/schemas/TimeBand"
              },
              "title": "Timeband",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "tagId",
            "required": false,
            "schema": {
              "items": {
                "format": "uuid",
                "type": "string"
              },
              "title": "Tagid",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "genderBalanceCode",
            "required": false,
            "schema": {
              "items": {
                "$ref": "#/components/schemas/GenderBalanceCode"
              },
              "title": "Genderbalancecode",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/CircleSort",
              "default": "newest"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CircleCollectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "minLength": 1,
                      "title": "Code",
                      "type": "string"
                    },
                    "detail": {
                      "minLength": 1,
                      "title": "Detail",
                      "type": "string"
                    },
                    "errors": {
                      "anyOf": [
                        {
                          "items": {
                            "additionalProperties": false,
                            "properties": {
                              "code": {
                                "minLength": 1,
                                "title": "Code",
                                "type": "string"
                              },
                              "field": {
                                "minLength": 1,
                                "title": "Field",
                                "type": "string"
                              }
                            },
                            "required": [
                              "field",
                              "code"
                            ],
                            "title": "ProblemFieldError",
                            "type": "object"
                          },
                          "type": "array"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Errors"
                    },
                    "instance": {
                      "minLength": 1,
                      "title": "Instance",
                      "type": "string"
                    },
                    "requestId": {
                      "format": "uuid",
                      "title": "Requestid",
                      "type": "string"
                    },
                    "status": {
                      "maximum": 599,
                      "minimum": 100,
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "minLength": 1,
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "default": "about:blank",
                      "minLength": 1,
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "code",
                    "requestId"
                  ],
                  "title": "ProblemDetails",
                  "type": "object"
                }
              }
            },
            "description": "Invalid pagination cursor"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "minLength": 1,
                      "title": "Code",
                      "type": "string"
                    },
                    "detail": {
                      "minLength": 1,
                      "title": "Detail",
                      "type": "string"
                    },
                    "errors": {
                      "anyOf": [
                        {
                          "items": {
                            "additionalProperties": false,
                            "properties": {
                              "code": {
                                "minLength": 1,
                                "title": "Code",
                                "type": "string"
                              },
                              "field": {
                                "minLength": 1,
                                "title": "Field",
                                "type": "string"
                              }
                            },
                            "required": [
                              "field",
                              "code"
                            ],
                            "title": "ProblemFieldError",
                            "type": "object"
                          },
                          "type": "array"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Errors"
                    },
                    "instance": {
                      "minLength": 1,
                      "title": "Instance",
                      "type": "string"
                    },
                    "requestId": {
                      "format": "uuid",
                      "title": "Requestid",
                      "type": "string"
                    },
                    "status": {
                      "maximum": 599,
                      "minimum": 100,
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "minLength": 1,
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "default": "about:blank",
                      "minLength": 1,
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "code",
                    "requestId"
                  ],
                  "title": "ProblemDetails",
                  "type": "object"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "minLength": 1,
                      "title": "Code",
                      "type": "string"
                    },
                    "detail": {
                      "minLength": 1,
                      "title": "Detail",
                      "type": "string"
                    },
                    "errors": {
                      "anyOf": [
                        {
                          "items": {
                            "additionalProperties": false,
                            "properties": {
                              "code": {
                                "minLength": 1,
                                "title": "Code",
                                "type": "string"
                              },
                              "field": {
                                "minLength": 1,
                                "title": "Field",
                                "type": "string"
                              }
                            },
                            "required": [
                              "field",
                              "code"
                            ],
                            "title": "ProblemFieldError",
                            "type": "object"
                          },
                          "type": "array"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Errors"
                    },
                    "instance": {
                      "minLength": 1,
                      "title": "Instance",
                      "type": "string"
                    },
                    "requestId": {
                      "format": "uuid",
                      "title": "Requestid",
                      "type": "string"
                    },
                    "status": {
                      "maximum": 599,
                      "minimum": 100,
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "minLength": 1,
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "default": "about:blank",
                      "minLength": 1,
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "code",
                    "requestId"
                  ],
                  "title": "ProblemDetails",
                  "type": "object"
                }
              }
            },
            "description": "Unexpected server error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "minLength": 1,
                      "title": "Code",
                      "type": "string"
                    },
                    "detail": {
                      "minLength": 1,
                      "title": "Detail",
                      "type": "string"
                    },
                    "errors": {
                      "anyOf": [
                        {
                          "items": {
                            "additionalProperties": false,
                            "properties": {
                              "code": {
                                "minLength": 1,
                                "title": "Code",
                                "type": "string"
                              },
                              "field": {
                                "minLength": 1,
                                "title": "Field",
                                "type": "string"
                              }
                            },
                            "required": [
                              "field",
                              "code"
                            ],
                            "title": "ProblemFieldError",
                            "type": "object"
                          },
                          "type": "array"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Errors"
                    },
                    "instance": {
                      "minLength": 1,
                      "title": "Instance",
                      "type": "string"
                    },
                    "requestId": {
                      "format": "uuid",
                      "title": "Requestid",
                      "type": "string"
                    },
                    "status": {
                      "maximum": 599,
                      "minimum": 100,
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "minLength": 1,
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "default": "about:blank",
                      "minLength": 1,
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "code",
                    "requestId"
                  ],
                  "title": "ProblemDetails",
                  "type": "object"
                }
              }
            },
            "description": "Service temporarily unavailable"
          }
        },
        "summary": "List public Circles",
        "tags": [
          "API v1"
        ]
      }
    },
...
  }
}